Damage system

+++ src/components/HomingComponent.h
@@ -10,13 +10,13 @@
 class HomingComponent : public wolf::BaseComponent
 {
 public:
-    HomingComponent(wolf::GameObject * p_target, float p_turning_cap, int p_delay = 0);
+    HomingComponent(wolf::GameObject * p_target, float p_turning_cap, float p_delay = 0);
     virtual ~HomingComponent();
 
     void Update(float p_delta);
 private:
     wolf::GameObject* m_pTarget = nullptr;
-    int m_iUpdateDelay = 0;
-    int m_iUpdateDelayCounter = 0;
+    float m_iUpdateDelay = 0.0f;
+    float m_iUpdateDelayCounter = 0.0f;
     float m_fTurningCapRad = 0.0f;
 };
